update: move errhandling to middleware
